+++ tests/Unit/DSC_DnsServerAddress.Tests.ps1
@@ -32,6 +32,12 @@ Invoke-TestSetup
 try
 {
     InModuleScope $script:dscResourceName {
+
+        $commonLocalizedData = Get-LocalizedData `
+            -ResourceName 'NetworkingDsc.Common' `
+            -ScriptRoot (Join-Path -Path $PSScriptRoot `
+                -ChildPath '..\..\source\Modules\NetworkingDsc.Common\')
+
         Describe 'DSC_DnsServerAddress\Get-TargetResource' -Tag 'Get' {
             Context 'Test IPv4' {
                 Context 'Invoking with an IPv4 address and one address is currently set' {
@@ -608,7 +614,7 @@ try
                     }
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressFormatError -f $assertResourcePropertySplat.Address) `
+                        -Message ($commonLocalizedData.AddressFormatError -f $assertResourcePropertySplat.Address) `
                         -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @assertResourcePropertySplat } | Should -Throw $ErrorRecord
@@ -625,7 +631,7 @@ try
                     }
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv4MismatchError -f $assertResourcePropertySplat.Address, $assertResourcePropertySplat.AddressFamily) `
+                        -Message ($commonLocalizedData.AddressIPv4MismatchError -f $assertResourcePropertySplat.Address, $assertResourcePropertySplat.AddressFamily) `
                         -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @assertResourcePropertySplat } | Should -Throw $ErrorRecord
@@ -642,7 +648,7 @@ try
                     }
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv6MismatchError -f $assertResourcePropertySplat.Address, $assertResourcePropertySplat.AddressFamily) `
+                        -Message ($commonLocalizedData.AddressIPv6MismatchError -f $assertResourcePropertySplat.Address, $assertResourcePropertySplat.AddressFamily) `
                         -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @assertResourcePropertySplat } | Should -Throw $ErrorRecord