+++ tests/Unit/DSC_Route.Tests.ps1
@@ -66,6 +66,11 @@ try
             PreferredLifetime = ([Timespan]::FromSeconds($testRoute.PreferredLifetime))
         }
 
+        $commonLocalizedData = Get-LocalizedData `
+            -ResourceName 'NetworkingDsc.Common' `
+            -ScriptRoot (Join-Path -Path $PSScriptRoot `
+                -ChildPath '..\..\source\Modules\NetworkingDsc.Common\')
+
         Describe 'DSC_Route\Get-TargetResource' -Tag 'Get' {
             Context 'Route does not exist' {
                 Mock -CommandName Get-NetRoute
@@ -390,8 +395,8 @@ try
                     $Splat.AddressFamily = 'IPv4'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressFormatError -f '10.0.300.0') `
-                        -ArgumentName 'DestinationPrefix'
+                        -Message ($commonLocalizedData.AddressFormatError -f '10.0.300.0') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -407,8 +412,8 @@ try
                     $Splat.AddressFamily = 'IPv6'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressFormatError -f 'fe8x::') `
-                        -ArgumentName 'DestinationPrefix'
+                        -Message ($commonLocalizedData.AddressFormatError -f 'fe8x::') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -424,8 +429,8 @@ try
                     $Splat.AddressFamily = 'IPv4'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv6MismatchError -f 'fe80::', 'IPv4') `
-                        -ArgumentName 'DestinationPrefix'
+                        -Message ($commonLocalizedData.AddressIPv6MismatchError -f 'fe80::', 'IPv4') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -441,8 +446,8 @@ try
                     $Splat.AddressFamily = 'IPv6'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv4MismatchError -f '10.0.0.0', 'IPv6') `
-                        -ArgumentName 'DestinationPrefix'
+                        -Message ($commonLocalizedData.AddressIPv4MismatchError -f '10.0.0.0', 'IPv6') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -458,8 +463,8 @@ try
                     $Splat.AddressFamily = 'IPv4'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressFormatError -f '10.0.300.0') `
-                        -ArgumentName 'NextHop'
+                        -Message ($commonLocalizedData.AddressFormatError -f '10.0.300.0') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -475,8 +480,8 @@ try
                     $Splat.AddressFamily = 'IPv6'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressFormatError -f 'fe9x::') `
-                        -ArgumentName 'NextHop'
+                        -Message ($commonLocalizedData.AddressFormatError -f 'fe9x::') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -492,8 +497,8 @@ try
                     $Splat.AddressFamily = 'IPv4'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv6MismatchError -f 'fe90::', 'IPv4') `
-                        -ArgumentName 'NextHop'
+                        -Message ($commonLocalizedData.AddressIPv6MismatchError -f 'fe90::', 'IPv4') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }
@@ -509,8 +514,8 @@ try
                     $Splat.AddressFamily = 'IPv6'
 
                     $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($script:localizedData.AddressIPv4MismatchError -f '10.0.1.0', 'IPv6') `
-                        -ArgumentName 'NextHop'
+                        -Message ($commonLocalizedData.AddressIPv4MismatchError -f '10.0.1.0', 'IPv6') `
+                        -ArgumentName 'Address'
 
                     { Assert-ResourceProperty @Splat } | Should -Throw $errorRecord
                 }